BUG#26580064: CONFUSING ERROR MESSAGE IN REPLICATION WHEN CONNECTING WITH MASTER
When replication connection fails due to some error, the original
error message is overwritten with a generic message, both in SHOW
SLAVE STATUS and in performance_schema.replication_connection_status.
Therefore, it is hard to find the reason for the connection failures.

Fixed by appending the original message to the end of the generic one.

RB: 21810