Bug #29550527 CONDITIONAL JUMP... IN UPGRADE TESTS
Problem:
Several upgrade tests are failing with the following symptom in Valgrind:

__strcmp_sse42 (vg_replace_strmem.c:851)

The problem happens if the tablename is not formatted to include the database name. The call to dd_parse_tbl_name returns false if the table name does not have the database name and hence, the buffer holding the database name remains uninitialized. The tables in question here are SDI tables which do not have the format dbname/tablename causing dd_parse_tbl_name to return false with an uninitialized db_buf.

This uninitialized string causes the valgrind enabled tests to fail as reported in the bug.

Solution:
Initialize the db_buff array before passing to dd_parse_tbl_name.

RB: 21807
Reviewed By: Annamalai Gurusami (annamalai.gurusami@oracle.com)