Bug#28934315 CASE STATEMENT USE IN CONDITIONAL WITH SUB_PART INDEX REGRESSION IN 8.0.
If prefix index is used for the REF access and the function is used as a value for the
key comparison, result of the function can be longer than prefix index length. In this
case the function result is truncated to the length of the prefix key when it's stored
to the key field. type_conversion_status_to_store_key() function returns
store_key::STORE_KEY_CONV and cp_buffer_from_ref() treats it as an error despite
THD::check_for_truncated_fields is set to CHECK_FIELD_IGNORE and
TABLE::set_no_row() is set. So no row is output.
Reviewed-by: Roy Lyseng <roy.lyseng@oracle.com>