Bug #23304911: ARCHIVE ENGINE - HEAP-USE-AFTER-FREE witn archive RB#21820
GIS blobs have an optimization that, when outlining them, avoids copying
the data and instead points to the source location used to set them.
This works well in the optimizer where these locations are mostly
ephemeral and a product of e.g. functions etc.
But REPAIR TABLE on archive tables uses a single buffer: it first uses
it as intermediary to read disk data (and copy them over to the Item
values). Then it reallocates the buffer (since it needs a bit of a bigger
buffer to write) and then writes data into it as a preparation to
write to disk. The re-allocation results in freeing the old buffer
and allocating a new one.

Thus for GIS fields that don't copy the actual data and just the pointer
to the incoming data the pointer points to a buffer that's freed. And
worse, can be overwritten by bad data.

Fixed by raising the flag so that GIS blobs are deep-copied.