Bug#29588413: UNNECESSARY VIRTUAL FUNCTIONS IN FIELD CLASSES
Removed some overrides of store() in Field_string, Field_varstring and
Field_bit_as_char. Their implementations simply forwarded the call to
the corresponding function in the parent class, so they could be
inherited instead. Two more store() overloads could be marked as final
in Field_bit because of this change.

Field::charset_for_protocol() was turned into a non-virtual function,
as it is not overridden by any subclass.

Change-Id: I38af8d1bff73f61c08d424f39c39fe619f9bde64