Bug #29587962 TESTBACKUP -N CONSISTENCYUNDERLOAD* TESTS REPORTED AS FAILED EVEN WHEN PASSING
Reviewed-by: Frazer Clement <frazer.clement@oracle.com>

The ConsistencyUnderLoad testcases in testBackup contain an
ndb_restore epoch restore which is allowed to fail. So the overall
result is that the test passes, but the test output contains a
test failure string. Due to this, autotest reports the test as failed.

This fix adds a workaround to avoid printing the test failure string
to the test log. A check is added to skip the epoch restore if the
ndb_apply_status table is not found.