BUG#29601546
This is the same problem as in BUG#29487340.

The problem in the fix of this bug is that it only sends
lcp_already_completed as calculated by the signal received last.

This means that the fix of that bug was dependent on signal order. To avoid
that we need to save the lcp_already_completed in a block variable until all
have responded to MASTER_LCPREQ. If any has completed the LCP, this variable
should be set.

Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>