Fix for "binding socket failed" in router component tests on Windows
On Windows sometimes there is a bunch of test failing with an error:
binding socket failed.
This is caused by the libevent not setting REUSEADDR on the socket it
uses on Windows to accept the requests in the http server.

This patch manually sets up a socket before passing it to the libevent for further processing.
REUSEADDR option is now set on all platforms.

RB: 21875
Reviewed-by: Jan Kneschke <jan.kneschke@oracle.com>