Bug#29622664 WITH_LOCK_ORDER DOES NOT BUILD, AS FLEX IS COMMENTED IN CMAKE
Currently, the WITH_LOCK_ORDER build is broken,
as some flex generated code is not invoked from the makefile.

Removing flex from the makefile was done as a temporary work around,
to allow to build the MySQL code base without flex installed,
as requiring flex even when not building WITH_LOCK_ORDER was too intrusive.

This fix is a rework on the makefile, around flex and bison,
related to the LOCK_ORDER feature, to have a permanent, less intrusive solution.

In particular:
- When building WITH_LOCK_ORDER, the flex tool must be available
- When building without LOCK_ORDER, the flex tool is *not* required.

In addition, the following changes are implemented:
- the cmake/flex.cmake file is abandoned, replaced by FindFLEX in CMake
- For sql/debug_lo_parser.yy, FindBISON is used instead of
  cmake/bison.cmake.
  Note that other existing uses of bison (SQL parser, hints parser)
  are unchanged.
- The LOCK_ORDER scanner (debug_lo_scanner.cc) and parser
  (debug_lo_parser.cc) are not included in the 'make dist' target.

LOCK_ORDER is considered a maintainer tool,
typically used when building from source,
and building it requires extra tooling like flex.

Also fix several compiler warnings:
 - Clang: missing 'override' attribution of misc virtual functions.
 - OsX: wrong printf format %ld for thread id arguments
 - 32bit build: wrong printf format %ld for size_t arguments

Change-Id: I152b444eb96082d9ec1b2ed1bfb7469dbcb69337