Bug#29492113 TABLE NOT ACCESSIBLE AFTER MOVING THE IBD FILE, WHEN THE DATABASE NAME HAS "/"
The problems happen because of incorrect parsing of dbname and tablename
if they have a "/" in their names. The server considers them as
directory separator and not part of the name itself.

This leads to situations where incorrect strings are considered as
dbname or filename leading to multiple issues.

Solution:
The dbname and filename are correctly parsed when converted to the
filesystem charset. However, in some cases, the name strings were
in the tablespace charset leading to the problems. The solution is to
use these names consistently in filesystem charset and convert them to
the tablespace charset as and when needed.

RB# 21921
Reviewer: Kevin Lewis (kevin.lewis@oracle.com)