WL#12956 Extend condition pushdown to any expression where 'used_tables' are available
Introduce usage of Item::used_tables() as the metric for deciding
when an Item expression is pushable or not. The used_tables map for
an Item expression contain the set of all tables referred by the expression.
Thus, if the 'used_tables' covers only (rows from) tables already being read
prior to 'this' table, the Item expression is effectively a 'const'
expression at this point in time.

The above used_table check will cover a super set of the current
cases handling:

 1) The 'Item' is a Field from previous table in the query plan.
    (Introduced in WL12686)
 2) Item is a constant expression (like -17, 1+2, concat(...))
 3) Item is a numeric or string literal (STRING_ITEM, INT_ITEM,
    REAL_ITEM, VARBIN_ITEM, DECIMAL_ITEM).
 4) Item is a 'CACHE_ITEM'

Thus, the more or less identical handling of these 4 special cases is now
combined into a single case of handling a 'used table covered the expression'.
In addition we get for free that any combination of the previous special cases
now also become pushable.

In some cases even entire predicate terms may be pre-evaluated to its
boolean true / false value.

In order to take advantage of the above improvements the 'rewrite' of IN and BETWEEN predicate
has also been refactored and partly reimplemented:

Previously the rewrite logic was splitt across 3 different places in
ndb_serializ...
 1) An early code section in ndb_serialize_cond setting up and
    checking the'left_hand_item'
 2) Another section after the one above handling the 'right side'
    of the operation: Setting up the 'expectation' for acceptable types
    and partly generating some of the serialized code.
3) At the end of ndb_serialize_cond: Clean up the rewrite context
   when last operand had been processed.

BETWEEN and IN rewrite was handled as a special case, with their
own rules for what to expect. Thus making it hard to take advantage
of other improvement in the condition pushdown code without having to
be maintained seperately.

As part of this patch,  the logic in the 3 steps above is collected into a common
code block. Furthermore, instead of doing code generation directly as
a special case, we completely transform the predicate by constructing
new Item_func_ge, Item_func_le, Item_func_eq predicates and let
ndb_serialize_cond() itself process the rewritten predicate terms.

Furthermore, we now also allow conditions inside a IS_NOT_NULL_COMPlimented trigger
 to be pushed. These conditions are intended to only be evaluated when the row
is not 'null complimented', which is exactly when a pushed condition
is evaluated. So it should be safe to push these predicates
whenever possible. The other variant of a IS_NOT_NULL_COMP trigger
is the FOUND_MATCH trigger. The condition associated with this trigger
is used when a NULL-extended row *is* created from the table, which is
something we should NOT push. (the row effectively does not exists)

Reviewed by: Martin Skold  <Martin.Skold@oracle.com>

Change-Id: I86808dec238c2a7834faf9c5245a67fdb105d76e