Bug#29585709 ndb_restore ignores old-style restore list (patch ver. 2)
When invoked as "ndb_restore --exclude-databases=mysql test t1", ndb_restore
would detect the "new-style" --exclude- argument and then ignore the
"old-style" restore list "test t1".

There is no reason --exclude-databases= and --exclude-tables= should be
incompatible with the old restore list, so do not ignore it if they are
used.

Change-Id: I3f835a92872fddeeddead155cef87776e2a4a947