Bug #29528188 NDB : TC FAILURE ON NDBASSERT(TRANSP->NDBAPIBLOCKREF == APIBLOCKREF)
NdbApiSignal objects are used to define signals for sending
to data nodes.

Each Ndb object maintains an internal pool of NdbApiSignal objects
which are reused.

When an NdbApiSignal object is allocated, it is initialised with the
block reference of the Ndb object passed to its constructor.

This block reference is then used every time the signal object is
used to send a signal.

---

The NdbApiSignal object is currently initialised using the
'theMyRef' member of the Ndb object, which is only initialised
when the node id of the API is known.  If an NdbApiSignal
object is intialised before the node id of the API is known
(e.g. before or during connection), then it may be set to
zero, triggering this problem.

This is solved by modifying the NdbApiSignal initialisation
to use the 'theNdbBlockNumber' variable instead, which is
initialised even before connecting, and should never be
zero.

Reviewed-by : Mikael Ronstrom <mikael.ronstrom@oracle.com>