BUG#29417503 MYSQL COULD HANG DURING SHUTDOWN BECAUSE OF ZOMBIE IN MYSQL_TRX_LIST
Proper and efficient mechanism to check if threads are gone is delivered.
Based on that the code is fixed for all background threads that are started
in InnoDB.

Also, shutdown in InnoDB was not precise enough regarding when the master
thread should exit. That is fixed - a new intermediate state of shutdown
is introduced, which is placed between CLEANUP and FLUSH_PHASE and marks
the point since which master should be closed.

RB: 21627