BUG#29386503 STOP SLAVE SQL_THREAD DEADLOCKS IF DONE WHILE HOLDING LOCK INSTANCE FOR BACKUP
Description
-----------
Server hangs if a `STOP SLAVE` statement is issued after a `LOCK INSTANCE FOR
BACKUP` and while an applier worker is executing a DDL statement.

Analysis
--------
If a DDL statement is applied after a `LOCK INSTANCE FOR BACKUP` is issued, it
will make the worker thread to block on the `BACKUP` lock acquisition.
Executing, at this point, `STOP SLAVE` will make the connection thread to wait
for every worker to stop. Since there is a worker waiting on the `BACKUP` lock,
we reach a dead-lock.

Fix
---
At the beginning of `STOP SLAVE` statement execution, try to acquire the
`BACKUP` lock. If it's possible to acquire it, do it and proceed. If it's not
possible to acquire, issue an error to the client connection.

Reviewed-by: Daogang Qu <bill.qu@oracle.com>
Reviewed-by: Neha Kumari <neha.n.kumari@oracle.com>
ReviewBoard: 21759