WL#11734 Ndb Inplace Rename Column
Added testcases to test inplace column rename and verifying
that the new column name is seen in all interfaces. If replication is setup
the binlog content as well as slave changes should be verified.
Error handling of illegal column name changes should be verified.
Backup/restore and system restarts from logs should verify that the column
name changes are persistent.

Reviewed-by: Maitrayi Sabaratnam <maitrayi.sabaratnam@oracle.com>