Bug#29639381 SCHEMA OPERATIONS NOT ABORTED
The communication between schema dist client and coordinator is done
with NDB_SCHEMA_OBJECT as well as writing rows to the ndb_schema table
in NDB. This opens up for different race conditions between when the
schema operation is registered and coordinator is notified of schema
operation.

- Fix problem where coordinator didn't abort active schema operations
when binlog thread was restarted.
- Fix problem where registered schema operations was not aborted
properly.
- Fix problem where client didn't detect schema distribution
not ready correctly.
- Fix problem where client detected that it was killed and just
left without marking the schema operation as failed.
- Fix problem where "op" in NDB_SHARE was accessed without proper locks
held.
 - Remove usage of ndb_schema_share global pointer and instead detect if
schema distribution is ready by checking if "op" in NDB_SHARE for
mysql.ndb_schema has been created.

Change-Id: Ifbde936e1e29aa8d4b69185258677f0dbeb7d9bd