WL#12009 - Redo-log Archiving WL#12329 - Redo-log Archiving Service API WL#12348 - Redo-log Archiving <antonym of endangerment>
Post-push fix. Unittest merge_innodb_tests-t failed.
If the initialization of the redo log archive feature failed
(registration of privilege or UDFs failed), it deinitializes
itself. But InnoDB starts up normally. When innodb is
deinitialized, it does also deinitialize the redo log
archive feature. That feature was not prepared to be
deinitialized twice.

Fix initialization/deinitialization so, that it keeps track
of its state.

RB#21822