Bug #29638626 NDB_RESTORE_GCP FAILURE IN PB2
From: Frazer Clement <frazer.clement@oracle.com>
Reviewed-by: Priyanka Sangam <priyanka.sangam@oracle.com>

The testBackup framework handles restores by copying backup files
from the remote host to the local host prior to running ndb_restore.
However, for platforms like linux-valgrind and linux-gcov, the
backup files are located on the localhost. The use of ssh and scp
on the localhost results in RSA host key identification errors.

The backup file-copy has been modified to detect the case where
the backups are located on the localhost. In this case, scp and ssh
are replaced by shell commands which can be executed locally.