BUG#29363787 / BUG#94356: Fix time granularity in dump thread
This patch fixes the precision of the parties involved in
calculating whether a heartbeat is needed to be sent to
the slave or not. They are now all done using nanoseconds,
even though the precision of the MASTER_HEARTBEAT_PERIOD
(for the CHANGE MASTER command) is set to seconds. We
default to nanoseconds, since that is the precision used
by the slave when connecting to the master. I.e., slave
first converts to nanoseconds and then interacts with
the master.

RB#21804