WL#13126 Dynamic linking of Protobuf
Build shared libraries for bundled protobuf.

The main changes of the patch are about installation, i.e. where to
put the new libraries when doing 'make install'. See changes in
install_layout.cmake and misc RPM/DEB packages. For RPM/DEB packages
it is important to *not* install the shared protobuf libraries in a
public directory (i.e. not together with libmysqlclient.so) since the
protobuf libraries are private to the mysql server. For other UNIX
installation layouts, the actual placement isn't really important.

Add a new cmake variable INSTALL_PRIV_LIBDIR to control location of
libprotobuf(-lite).so

Other changes:

In directories which use protobuf, add this to enable the necessary
dllexport/dllimport declarations:
IF(MSVC)
  ADD_DEFINITIONS(-DPROTOBUF_USE_DLLS)
ENDIF()

This is now done in
plugin/x/CMakeLists.txt
router/CMakeLists.txt

Targets which link with libprotobuf(-lite) directly or indirectly,
must be patched for INSTALL_RPATH on all non-windows platforms.

These targets are now
 - mysqlxtest  in plugin/x/tests/driver/CMakeLists.txt
 - mysqld      in sql/CMakeLists.txt

Most of the complexity: INSTALL_RPATH and @loader_path and libprotobuf.ver
is due to INSTALL_DEBUG_TARGET which is really broken.

Change-Id: Id7113b9dba35c1e0b65e4084ff7c35f5ebbd821c