Bug#29633473 FAILURE OF DOWNGARDE TESTS FROM TRUNK
Modifies upgrade and downgrade configuration files to use
mysql_native_password as default-authentication-plugin.

When downgrade tests are run from trunk to 7.4 or lower versions, the
upgrade database used to store atrt commands will be of the latest(trunk)
version. So, when the testcase is restarted, the test program of the
lower version fails to connect to the server of higher version with an
error saying Authentication plugin 'caching_sha2_password' cannot be loaded.
So, mysql_native_password is supported from mysql version 5.5 and above,
it is made default for these tests involving lower versions.

Approved by: Tiago Alves <tiago.alves@oracle.com>