Bug#25933891 - STATUS_BY_THREAD DOES NOT HAVE ANY DATA WHEN                THE THREAD POOL PLUGIN IS ENABLED
DESCRIPTION:
============
When the thread pool plugin is enabled, the
performance_schema.status_by_thread does not contain any
data.

ANALYSIS:
=========
Without thread pool, the THD is associated with the
instrumented thread while handling new incoming connections.
However, without the thread pool, the association of
THD to the instrumented thread was missing because of which
the data from some pfs tables could not be retrieved.

FIX:
====
Make sure that the thread is associated with the
instrumented thread so that the output of the table
status_by_thread is fetched correctly. The patch also fixes
a similar problem for the pfs table variables_by_thread.

RB: 21798