Bug#29276063 ASSERT IN CREATE_TMP_FIELD_FROM_ITEM() OR SEGFAULT IN CREATE_TMP_FIELD()
Problem:

We hit an assert in trying to create a tmp table field for a ROW
constructor. The tmp table is created because the select expression
contains a window function.

In the repro case we have a ROW constructor comparison, where the
right side is a subquery, so it can contain only at most one row (it
has `substype' Item_subselect::SINGLEROW_SUBS), but it has two select
columns in our repro case, so it is not a scalar subquery (requires a
degree of 1).

This means the result type of Item_singlerow_subselect is ROW_RESULT,
for which create_tmp_field_from_item is not prepared. So, we should
never have attempted to call create_tmp_field_from_item for it. Why
does this happen?

split_sum_func2 is called in the presence of aggregates or window
functions. It will move aggregates out of the expression's item tree
and replace with with references. In the repro, this happens with the
left side of the comparison, since it contains a window function;
COUNT(*). So far, so good. Since the row comparison contains an
aggregate or window function, we also check the right side of the
comparison, which does not contain any window function or
aggregate. But it does contain a subquery, as mentioned.

In 5.8 (not in 5.7) split_sum_func2 *also* explains it will replace
scalar subqueries with references to it as mentioned in this piece of
comment:

"Item_subselect can be added to "fields" only if it's a scalar subquery;
indeed a subquery of another type is wrapped in Item_in_optimizer at this
stage, so when splitting Item_in_optimizer, if we added the underlying
Item_subselect to "fields" below it would be later evaluated by
copy_fields() (in tmp table processing), which would be incorrect as the
Item_subselect cannot be evaluated - as it must always be evaluated
through its parent Item_in_optimizer."

This is not true, in our case, we do not have an IN (subquery), an
hence no Item_in_optimizer wrapping the subquery, but that aside, we
do not have a scalar subquery either, so it should not be replaced by
a reference.

The problem lies in how split_sum_func2 identifies a scalar
subqery. It does indeed check for substype being equal to
Item_subselect::SINGLEROW_SUBS (i.e. that the subquery is limited to
max one row in the result set), but neglects to check whether the
subquery has degree 1. Hence, we can erroneously identify a subquery
as scalar even when it is not, and try to create a tmp field for it
(as in the repro). Note: if we replace the windowed aggregate with a
grouped aggregate in the repro, we also get a crash for them same
reason, a regression from 5.7.

Solution: add the predicate that degree be 1, so we identify a scalar
subquery correctly.

Question: In what cases do we need to create a reference to the scalar
subquery? Removing it makes this query from the test suite fail:

SELECT FIRST_VALUE((SELECT upper.j FROM t1 LIMIT 1))
    OVER (ORDER BY i) FROM t1 AS upper;

i.e. here the subquery is an argument to window function, since the
value of the scalar subquery must be written to the tmp table used by
the window function. What would happen if we have a ROW predicate with
a subquery as an argument to a window function? In this case, the =
comparison doesn't contain an aggregate or window reference, so
split_sum_func2 doesn't encounter the ROW constructor:

SELECT LAST_VALUE(ROW(1,1) = (SELECT 1, upper.j FROM t LIMIT 1))
    OVER (ORDER BY i)  FROM t1 AS upper;

So in conclusion, we only need split_sum_func2 to create references to
(always scalar) subqueries for the purpose of references from window
function. Added that to Doxygen. Verified that regression suite
doesn't need this except inside wf arguments.

A strangeness (misleading at least) is the reference created for the
references in split_sum_func2 is typed Item_aggregate_ref even when
the reference isn't (always) to an aggregate item. It should probably
be renamed.

Change-Id: Id04993fbe33cdac9bb779a5cde0304637271069f