Bug #28835179: INCONSISTENT RESULTSET FOR DIFFERENT EXECUTION PLANS
ISSUE:Inner tables of different semi join nests are getting
interleaved for MaterializeScan.

SOLUTION: Added a check so that inner tables of different semi
join nests are prevented from getting interleaved for
MaterializeScan.

The fix is partial backport of changes gone in 8.0 as part of
WL#9158 and is based on contribution patch.

Change-Id: I0c2f1d1afb426b92366087fbbd8552c1cd279cdb

Approved by Roy Lyseng