Bug#28196680 RPL_LOCK_BACKUP_BINARY_AND_RELAY_LOGS FAILS : MASTER-BIN.000005 NOT PURGED ERROR
Suppresses warnings thrown when attempting to purge logs which are still being
read by a mysql server thread.

Problem: Some replication tests fail sporadically with warning "File was not purged
because it was being read by thread". This warnings is thrown when one tries
to purge a log which is being read by a mysql thread, situation that can occur
in slow systems.

Fix: Suppress the warning.

RB: 21994
Reviewed-by: Deepthi Eranti Sreenivas <deepthi.e.s@oracle.com>
Reviewed-by: Sven Sandberg <sven.sandberg@oracle.com>