Bug#29621062: MERGE FIELD::SEND_TEXT() AND FIELD::SEND_BINARY()
Followup patch: Make some assertions stricter. Before the bug was
fixed, the metadata in Protocol_classic::field_types could not be
trusted, so the assertions had to be lenient. Now, the metadata
matches the actual data type that is sent on the protocol, so the
assertions can be made stricter.

Change-Id: I8a0f8c39b0f45465560a9533b6ebedb15452ff3f