Bug #29127203 VIRTUAL GENERATED COLUMN INDEX DATA INCONSISTENCY
PROBLEM
-------

Index defined on a virtual column whose base column was in a fk
relation was not getting updated. This is because while getting
the updated field information from the update vector of the parent
table we were comparing the column number of the base column (for
virtual column) in child table with the associated column number
in the parent table. There was a mismatch in this column number
because of which this update field information was skipped and
subsequently index was not getting updated.


FIX
---

The mismatch in column can happen because the associated column
position in parent table may differ. To overcome this we get the
position of column with reference to child table first and then
match it with the column of parent table.

Approved by: Bin su
RB: 21970