Bug#28988955 BACKPORT BUG#26709230 TO 5.6.
This is a backport of 'Bug#26709230 MTR UNABLE TO DETECT ERRORS
IN ERROR LOG GENERATED BY TEST SCRIPT'

Issue
=====
MTR scans error logs of all servers it spawns one by one on test
completion to check if their error logs contain un-suppressed
warnings or errors using using the check_warnings setup. The flag
set when there is a check_warnings failure is being overwritten
if subsequent iterations of check_warnings on other servers are
successful. This might mask failures as MTR reports the test run
passed.

Fix
===
MTR altered to report a check_warnings() failure as soon as it
finds one i.e., it does not proceed to scan the error log of
the next server.

Reviewed-by: Deepa Dixit <deepa.dixit@oracle.com> (RB#22023)