Bug#29248047: 5.7 AUDIT PLUGIN DOES NOT LOG WHO UNINSTALLS               THE AUDIT PLUGIN UNLIKE 5.6
Description: Audit record for "UNINSTALL PLUGIN audit_log"
is not logged in the audit log for mysql-5.7+. But in 5.6,
this record is logged.

Analysis: In 5.7, this issue got introduced by the fix for
Bug#23236404. Before the fix, audit plugins are acquired on
the first use and released after the execution of each
query. As part of the fix, audit plugins are acquired on the
first use and released when the connection ends.

Before the fix, 'mysql_audit_release()' was done before
'do_command()' which will ensure that the plugin is
released after execution of each query. This guaranteed
"UNINSTALL PLUGIN audit_log" query to be recorded in the
audit log since the 'plugin_deinitialize()' [as part of
'mysql_audit_release()' during plugin release] happens after
the logging of the UNINSTALL PLUGIN audit record.

But as part of the fix, 'mysql_audit_release()' is moved
inside 'Sql_cmd_uninstall_plugin::execute()'. Due to this,
audit log gets deinitialized as part of
"plugin_deinitialize()" during the execution of
'Sql_cmd_uninstall_plugin::execute()'. As a result, when the
execution reaches the point where the audit record for
"UNINSTALL PLUGIN audit_log" is logged, it is unable to log
the record since the plugin is already deinitialized.

Fix: 'mysql_audit_release()' is removed for UNINSTALL PLUGIN
command. This is to make sure that a thread that has started
as being audited should not be forced into not being audited.

RB#21815