Bug#29621062: MERGE FIELD::SEND_TEXT() AND FIELD::SEND_BINARY()
Post-push fix for a failure in funcs_1.storedproc.

The original bug fix removed the override of Item::send() in
Item_sp_variable. However, it turns out that it is still needed in
case the variable has a ZEROFILL attribute. Item::send() doesn't know
about ZEROFILL, so the call has to be forwarded to Item_field::send().

It is fixed by reintroducing the removed send() override. A comment is
added to explain that it is only needed for ZEROFILL.

Change-Id: I1529fe9e8a51e0118a71bc1c49ef759df71f5bd2