Bug #27538464 	MORE THAN 2G SYSTEM TABLESPACE (IBDATA) IS NOT CREATED WHEN INITIALIZING
In v8.0 Datafile::m_size was converted from a ulint to uint32_t,
which is what page_no_t was defined as. This caused the previous
static_cast to do the bit-shifting as a 32 bit page number instead
of as a 64 bit offset.  So the offset could not get any bigger than 2G.

Solution, cast to os_offset_t first, the bit-shift.

This change is suggested by Sunny Bains and committed by Kevin Lewis
Approved by Sunny Bains <sunny.bains@oracle.com>