Bug#29697670: OVERRIDING SERVER DEFAULT CHARSET/COLLATION AFFECTS UPGRADE
The collation_connection is used as collation for string literals in
SQL queries, and its default value is the server collation, which can
be set by submitting --collation-server as a command line option when
starting mysqld.

The thread executing server-side SQL statements is using this collation,
and this may affect the execution of SQL queries during upgrade of the
data dictionary tables.

To fix this, we explicitly set collation_connection and collation_server
for these threads to MYSQL_DEFAULT_COLLATION_NAME, which is now
utf8mb4_0900_ai_ci.

Change-Id: I8e80380b151ed96e2c91e92bdc3da64e77cc67d8